typing for beacon config reducers
